+++ Package.swift
@@ -1,7 +1,0 @@
-import PackageDescription
-
-let package = Package(
-    name: "DGSideNavigation",
-    dependencies: [
-    ]
-)