+++ Samples/SideNavigationControllerSample-iOS/SideNavigationControllerSample-iOS/LeftViewController.swift
@@ -1,0 +1,20 @@
+//
+//  LeftViewController.swift
+//  SideNavigationControllerSample-iOS
+//
+//  Created by Benoit BRIATTE on 25/02/2017.
+//  Copyright © 2017 Digipolitan. All rights reserved.
+//
+
+import UIKit
+
+class LeftViewController: UIViewController {
+
+    override var prefersStatusBarHidden: Bool {
+        return true
+    }
+
+    override var preferredStatusBarUpdateAnimation: UIStatusBarAnimation {
+        return .slide
+    }
+}