+++ Samples/SideNavigationControllerSample-iOS/SideNavigationControllerSample-iOS/RightViewController.swift
@@ -1,0 +1,13 @@
+//
+//  RightViewController.swift
+//  SideNavigationControllerSample-iOS
+//
+//  Created by Benoit BRIATTE on 13/03/2017.
+//  Copyright © 2017 Digipolitan. All rights reserved.
+//
+
+import UIKit
+
+class RightViewController: UIViewController {
+
+}