+++ Sources/UIViewController+SideNavigationController.swift
@@ -1,6 +1,6 @@
 //
 //  UIViewController+SideNavigationController.swift
-//  DGSideNavigation
+//  SideNavigationController
 //
 //  Created by Benoit BRIATTE on 13/03/2017.
 //  Copyright © 2017 Digipolitan. All rights reserved.
@@ -10,7 +10,7 @@ import UIKit
 
 public extension UIViewController {
 
-    public var sideNavigationController: SideNavigationController? {
+    var sideNavigationController: SideNavigationController? {
         var current = self
         while let parent = current.parent {
             if let side = parent as? SideNavigationController {