+++ Tests/SideNavigationControllerTests/SideNavigationControllerTests.swift
@@ -1,0 +1,13 @@
+import XCTest
+@testable import SideNavigationController
+
+class SideNavigationControllerTests: XCTestCase {
+
+    override func setUp() {
+        super.setUp()
+    }
+
+    override func tearDown() {
+        super.tearDown()
+    }
+}