+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Change Log
 
 All notable changes to this project will be documented in this file.
-`side-navigation-swift` adheres to [Semantic Versioning](http://semver.org/).
+`side-navigation-controller` adheres to [Semantic Versioning](http://semver.org/).
 ---
 
-## [1.0.1](https://github.com/Digipolitan/side-navigation-swift/releases/tag/v1.0.1)
+## [2.0.0](https://github.com/Digipolitan/side-navigation/releases/tag/v2.0.0)
+
+swift 5.0 support & renaming side-navigation-controller
+
+---
+
+## [1.0.1](https://github.com/Digipolitan/side-navigation-controller/releases/tag/v1.0.1)
 
 move the addChildViewController to call viewWillAppear with the side set